Refactor screenshotting process and deprecate Squeak[Form]Picture in favor of (Screnshot|Form)Scene and ScreenshotRecorder which provide better units of state and responsibility.
Motivation: SqueakPicture has always been an impricise and misleading design decision. It does not actually derive from Form or something like this but rather represented the order to create such a figure.

Thus I deprecated it and migrated large parts of its implementation to the new ScreenshotScene hierarchy which should provide a clearer responsibilty.
The new ScreenshotRecorder allows recording one or even many screenshots sharing one world setup. This can be useful for many similar screenshots.
Its specialization reenables debugging of a scene, that is opening instead of saving its output picture. To debug a ScreenshotRecorder, just insert #debugging after a class or instance of it.

However, the public interface of SqueakPicture has been kept as forwarding to the new interface, for compatibility.