fix: update paragraph to fit preceding screenshot and do-this-action
In the "Type the following text into the workspace:" the reader should type "Transcript showln: 'hello world'." into their workspace and execute "do it (d)" afterwards. This corresponds to Figure 1.10. Alas the following paragraph tells about sending "show: 'hello world'" and "cr" to the Transcript object which was the case in the 5.3 version of SBE.